+++ lib/archipel/api.rb
@@ -3,8 +3,9 @@ module Archipel
     module Internal
     end
 
-    def self.defaults login: nil, password: nil, server: nil, hypervisor: nil, ** kwargs
+    def self.defaults login: nil, password: nil, server: nil, hypervisor: nil, **kwargs
       @config = {login: login, password: password, server: server, hypervisor: hypervisor}
+      @config.merge! kwargs
     end
 
     def self.get_defaults