+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Version history
 
-## 1 @ 2014-02-14
+## 4 @ 2014-02-17
+
+- new VmApi calls: info, autostart, stop, force_stop
+
+## 3 @ 2014-02-16
 
 - new VmApi calls: start, create_disk, list_disks, remove_disk, grant_permissions, revoke_permissions, subscribe,
 define_vm, xml
 - new HypervisorApi calls: register_user, list_users, unregister_user, create_vm, list_vm, delete_vm, grant_permissions
+
+## previous versions
+
+yanked